+++ src/main/java/fr/epsi/mspr/msprapi/repository/UserRepository.java
@@ -1,15 +1,14 @@
 package fr.epsi.mspr.msprapi.repository;
 
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import fr.epsi.mspr.msprapi.entities.User;
 
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
 	Optional<User> findByName(String name);
